requirements.txt
Django==2.2.1
django-cors-headers==3.0.2
PyJWT==1.7.1
pytz==2019.1
PyYAML==5.1
sqlparse==0.3.0
